Compiling successfully with no errors.
-- Image alt tags added

-- I removed fake commented out functionality and the imports for it,
because it was not even laid out the way a modern reduxized application
would be. This is definitely just my opinion, but if a component has no
logic and is just a view I feel it's worse to keep malformed half logic
within it than to show it as what is is for future developers.