Fixed final fixable error!!!
-- There is another warning in console from React concerning the prop
type within routers. Upon further research this seems to have to do with
the syntax of our entire application, and the libraries we have. It is
not fixable without completely changing every component, and it's only a
warning, so I'm not sweating it! Give it a try, it's great. No warnings!